tools/host_gui_hid/requirements.txt
hidapi==0.14.0.post4
